+++ crates/rust-release/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## 3.0.0 - 2026-09-11
 
+## 3.0.0 - 2026-09-11
+
 ### Added
 
 - Added `RustRelease::version_mut` which returns an exclusive reference to version of a release